changed 2 decimals to flaot

+++ server/migrations/20200614044545-create-expense.js
@@ -15,7 +15,7 @@ module.exports = {
         type: Sequelize.STRING
       },
       amount: {
-        type: Sequelize.DECIMAL(10,2),
+        type: Sequelize.FLOAT
       },
       description: {
         type: Sequelize.STRING
